backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
chromadb==0.4.15
pypdf==3.17.0
python-docx==1.1.0
markdown==3.5.1
rank-bm25==0.2.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
unstructured[pdf]==0.11.2
pillow==10.1.0